fix: develop브랜치 병합 충돌 해결 및 삭제 권한 로직 추가

+++ project/src/main/java/com/edison/project/domain/label/service/LabelCommandServiceImpl.java
@@ -88,9 +88,14 @@ public class LabelCommandServiceImpl implements LabelCommandService {
 
     @Override
     @Transactional
-    public void deleteLabel(Long labelId) {
+    public void deleteLabel(Long labelId, Long memberId) {
         Label label = labelRepository.findById(labelId)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.LABELS_NOT_FOUND));
+
+        // 삭제 권한 확인
+        if (!label.getMember().getMemberId().equals(memberId)) {
+            throw new GeneralException(ErrorStatus._UNAUTHORIZED);
+        }
 
         labelRepository.deleteById(label.getLabelId());
     }
